+++ druid-spring-boot-starter/src/main/java/com/alibaba/druid/spring/boot/autoconfigure/DruidDataSourceAutoConfigure.java
@@ -15,62 +15,47 @@
  */
 package com.alibaba.druid.spring.boot.autoconfigure;
 
+import javax.sql.DataSource;
+
 import com.alibaba.druid.pool.DruidDataSource;
-import com.alibaba.druid.spring.boot.autoconfigure.metadata.DruidDataSourcePoolMetadata;
+import com.alibaba.druid.spring.boot.autoconfigure.actuator.ActuatorConfiguration;
 import com.alibaba.druid.spring.boot.autoconfigure.properties.DruidStatProperties;
 import com.alibaba.druid.spring.boot.autoconfigure.stat.DruidFilterConfiguration;
 import com.alibaba.druid.spring.boot.autoconfigure.stat.DruidSpringAopConfiguration;
 import com.alibaba.druid.spring.boot.autoconfigure.stat.DruidStatViewServletConfiguration;
 import com.alibaba.druid.spring.boot.autoconfigure.stat.DruidWebStatFilterConfiguration;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties;
-import org.springframework.boot.autoconfigure.jdbc.metadata.DataSourcePoolMetadata;
-import org.springframework.boot.autoconfigure.jdbc.metadata.DataSourcePoolMetadataProvider;
-import org.springframework.boot.autoconfigure.jdbc.metadata.DataSourcePoolMetadataProvidersConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-
-import javax.sql.DataSource;
-
 /**
  * @author lihengming [89921218@qq.com]
  */
 @Configuration
-@ConditionalOnClass(com.alibaba.druid.pool.DruidDataSource.class)
+@ConditionalOnClass(DruidDataSource.class)
 @AutoConfigureBefore(DataSourceAutoConfiguration.class)
 @EnableConfigurationProperties({DruidStatProperties.class, DataSourceProperties.class})
 @Import({DruidSpringAopConfiguration.class,
-        DruidStatViewServletConfiguration.class,
-        DruidWebStatFilterConfiguration.class,
-        DruidFilterConfiguration.class})
+    DruidStatViewServletConfiguration.class,
+    DruidWebStatFilterConfiguration.class,
+    DruidFilterConfiguration.class,
+    ActuatorConfiguration.class})
 public class DruidDataSourceAutoConfigure {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(DruidDataSourceAutoConfigure.class);
 
     @Bean
     @ConditionalOnMissingBean
     public DataSource dataSource() {
+        LOGGER.info("Init DruidDataSource");
         return new DruidDataSourceWrapper();
-    }
-
-    /**
-     * Register the {@link DataSourcePoolMetadataProvider} instances to support DataSource metrics.
-     *
-     * @see DataSourcePoolMetadataProvidersConfiguration
-     */
-    @Bean
-    public DataSourcePoolMetadataProvider druidDataSourcePoolMetadataProvider() {
-        return new DataSourcePoolMetadataProvider() {
-            @Override
-            public DataSourcePoolMetadata getDataSourcePoolMetadata(DataSource dataSource) {
-                if (dataSource instanceof DruidDataSource) {
-                    return new DruidDataSourcePoolMetadata((DruidDataSource) dataSource);
-                }
-                return null;
-            }
-        };
     }
 }
