+++ druid-spring-boot-starter/src/main/java/com/alibaba/druid/spring/boot/autoconfigure/actuator/metadata/DruidDataSourcePoolMetadata.java
@@ -1,12 +1,26 @@
-package com.alibaba.druid.spring.boot.autoconfigure.metadata;
+/*
+ * Copyright 1999-2017 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.druid.spring.boot.autoconfigure.actuator.metadata;
 
 import com.alibaba.druid.pool.DruidDataSource;
+
 import org.springframework.boot.autoconfigure.jdbc.metadata.AbstractDataSourcePoolMetadata;
 import org.springframework.boot.autoconfigure.jdbc.metadata.DataSourcePoolMetadata;
 
 /**
- * {@link DataSourcePoolMetadata} for a Druid DataSource.
- *
  * @author lihengming [89921218@qq.com]
  */
 public class DruidDataSourcePoolMetadata extends AbstractDataSourcePoolMetadata<DruidDataSource> {
